Remove the useReservation Hook that is integrated into the useRooms
