+++ src/components/rooms_list.tsx
@@ -1,8 +1,7 @@
 import { styled } from '@stitches/react'
 import React, { FC } from 'react'
 import { RoomType } from '../atoms/rooms'
-import { RoomReservationStatusType } from '../hooks/useReservation'
-import useRooms from '../hooks/useRooms'
+import useRooms, { RoomReservationStatusType } from '../hooks/useRooms'
 import useTranslator from '../hooks/useTranslator'
 import { Floor } from './floor'
 import { RoomsListItem } from './rooms_list_item'
